Refactor CmsService and settings classes
Refactored `CmsService` to include `CmsSettings` dependency and updated `GetLatestArticles` to use configurable `HomeShownArticles`. Renamed `SearchSettings` to `CmsSettings` and adjusted properties. Renamed `CmsSettings` to `SearchSettings` in `SearchSettings.cs` with corresponding property updates.
