+++ ZucchiniMVC/Infrastructure/Config/CmsSettings.cs
@@ -1,9 +1,12 @@
 ﻿namespace Zucchinimvc.Infrastructure.Config
 {
-    public class SearchSettings
+    public class CmsSettings
     {
         public string BaseUrl { get; set; } = string.Empty;
-        public string ApiKey { get; set; } = string.Empty;
-        public string IndexName { get; set; } = string.Empty;
+        public string Token { get; set; } = string.Empty;
+        public int HomeShownArticles { get; set; } = int.MaxValue;
+
+
     }
+
 }
