+++ ZucchiniMVC/Infrastructure/Config/SearchSettings.cs
@@ -1,8 +1,10 @@
 ﻿namespace Zucchinimvc.Infrastructure.Config
 {
-    public class CmsSettings
+    public class SearchSettings
     {
         public string BaseUrl { get; set; } = string.Empty;
-        public string Token { get; set; } = string.Empty;
+        public string ApiKey { get; set; } = string.Empty;
+        public string IndexName { get; set; } = string.Empty;
     }
+
 }
